refactor controllers, and providers. Separated by classes
