+++ src/main/java/com/katalogus/project/controller/EloadasController.java
@@ -2,6 +2,7 @@ package com.katalogus.project.controller;
 
 import com.katalogus.project.entity.Eloadas;
 import com.katalogus.project.service.ClassesProvider;
+import com.katalogus.project.service.EloadasProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,11 +12,11 @@ import org.springframework.web.bind.annotation.*;
 public class EloadasController {
 
     @Autowired
-    ClassesProvider classesProvider;
+    EloadasProvider eloadasProvider;
 
     @PostMapping("/add")
     public ResponseEntity saveNewEloadas(@RequestBody Eloadas eloadas) {
-        Boolean successful = classesProvider.saveNewEloadas(eloadas);
+        Boolean successful = eloadasProvider.saveNewEloadas(eloadas);
         if (successful) {
             return ResponseEntity.ok("New Eloadas created successfully");
         } else {
@@ -25,7 +26,7 @@ public class EloadasController {
 
     @PutMapping("/{eloadas_id}/update")
     public ResponseEntity updateEloadasById(@RequestBody Eloadas eloadas, @PathVariable("eloadas_id") Long eloadasId) {
-        Boolean successful = classesProvider.updateEloadasById(eloadas, eloadasId);
+        Boolean successful = eloadasProvider.updateEloadasById(eloadas, eloadasId);
         if (successful) {
             return ResponseEntity.ok("Eloadas updated successfully");
         } else {
@@ -35,6 +36,6 @@ public class EloadasController {
 
     @DeleteMapping("/{eloadas_id}/delete")
     public void deleteEloadasById(@PathVariable("eloadas_id") Long eloadasId) {
-        classesProvider.deleteEloadasById(eloadasId);
+        eloadasProvider.deleteEloadasById(eloadasId);
     }
 }
