+++ src/main/java/com/katalogus/project/controller/GyakorlatController.java
@@ -2,6 +2,7 @@ package com.katalogus.project.controller;
 
 import com.katalogus.project.entity.Gyakorlat;
 import com.katalogus.project.service.ClassesProvider;
+import com.katalogus.project.service.GyakorlatProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,11 +12,11 @@ import org.springframework.web.bind.annotation.*;
 public class GyakorlatController {
 
     @Autowired
-    ClassesProvider classesProvider;
+    GyakorlatProvider gyakorlatProvider;
 
     @PostMapping("/add")
     public ResponseEntity saveNewGyakorlat(@RequestBody Gyakorlat gyakorlat) {
-        Boolean successful = classesProvider.saveNewGyakorlat(gyakorlat);
+        Boolean successful = gyakorlatProvider.saveNewGyakorlat(gyakorlat);
         if (successful) {
             return ResponseEntity.ok("New Gyakorlat created successfully");
         } else {
@@ -25,7 +26,7 @@ public class GyakorlatController {
 
     @PutMapping("/{gyakorlat_id}/update")
     public ResponseEntity updateGyakorlatById(@RequestBody Gyakorlat gyakorlat, @PathVariable("gyakorlat_id") Long gyakorlatId) {
-        Boolean successful = classesProvider.updateGyakorlatById(gyakorlat, gyakorlatId);
+        Boolean successful = gyakorlatProvider.updateGyakorlatById(gyakorlat, gyakorlatId);
         if (successful) {
             return ResponseEntity.ok("Gyakorlat updated successfully");
         } else {
@@ -35,6 +36,6 @@ public class GyakorlatController {
 
     @DeleteMapping("/{gyakorlat_id}/delete")
     public void deleteGyakorlatById(@PathVariable("gyakorlat_id") Long gyakorlatId) {
-        classesProvider.deleteGyakorlatById(gyakorlatId);
+        gyakorlatProvider.deleteGyakorlatById(gyakorlatId);
     }
 }
